move HP and Energy
