+++ arptypes/src/types.rs
@@ -1,3 +1,5 @@
+use derive_more::{Add, Div, Mul, Sub};
+use num::Saturating;
 use serde::{Deserialize, Serialize};
 use ts_rs::TS;
 use uom::si::length::{centimeter, meter};
@@ -22,8 +24,51 @@ pub fn i64meter<T: Into<i64>>(v: T) -> i64units::Length { i64units::Length::new:
 
 pub fn up_length(v: u32units::Length) -> i64units::Length { i64cm(v.get::<centimeter>()) }
 
-
 #[derive(Clone, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize, TS)]
 pub struct PlayerID(pub String);
 
+#[derive(
+  Add,
+  Sub,
+  Mul,
+  Div,
+  Clone,
+  Copy,
+  Eq,
+  PartialEq,
+  Ord,
+  PartialOrd,
+  Debug,
+  Hash,
+  Serialize,
+  Deserialize,
+  TS,
+)]
+pub struct HP(pub u8);
+impl Saturating for HP {
+  fn saturating_add(self, other: Self) -> Self { HP(self.0.saturating_add(other.0)) }
+  fn saturating_sub(self, other: Self) -> Self { HP(self.0.saturating_sub(other.0)) }
+}
 
+#[derive(
+  Add,
+  Sub,
+  Mul,
+  Div,
+  Clone,
+  Copy,
+  Eq,
+  PartialEq,
+  Ord,
+  PartialOrd,
+  Debug,
+  Hash,
+  Serialize,
+  Deserialize,
+  TS,
+)]
+pub struct Energy(pub u8);
+impl Saturating for Energy {
+  fn saturating_add(self, other: Self) -> Self { Energy(self.0.saturating_add(other.0)) }
+  fn saturating_sub(self, other: Self) -> Self { Energy(self.0.saturating_sub(other.0)) }
+}
